Add custom path option for single file reports.
This adds an option to override the path that single file reports are written to. The reason I need this feature is so I can have my report at "coverage/lcov.info" which is the only path supported by the Atom plugin lcov-info.

Also updated the "report generated" message so single files reports will just list that file's path, and per-file reports will give the path to the directory as before.